+++ src/kernel/chainstatemanager_opts.h
@@ -34,6 +34,7 @@ struct ChainstateManagerNotificationCallbacks {
     const std::function<void(SynchronizationState state, CBlockIndex* index)> notify_block_tip;
     const std::function<void(SynchronizationState state, int64_t height, int64_t timestamp, bool presync)> notify_header_tip;
     const std::function<void(const std::string& title, int nProgress, bool resume_possible)> show_progress;
+    const std::function<void(const bilingual_str& warning)> do_warning;
 };
 
 /**
